podcast-pro-plus/requirements.txt
fastapi[all]
python-dotenv[cli]
psycopg2-binary==2.9.9
pydantic-settings
pydub
thefuzz[speedup]
elevenlabs
requests
pytz
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
Authlib
feedparser
celery
Pillow
stripe
google-cloud-speech
google-generativeai
assemblyai
google-cloud-texttospeech
email-validator
sentry-sdk
slowapi
google-cloud-storage>=2.18.0
google-cloud-tasks>=2.16.5
python-multipart>=0.0.9
gunicorn==21.2.0
